% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_tester.R
\name{accuracy_tester_classification}
\alias{accuracy_tester_classification}
\title{Generate an accuracy tester function for classification}
\usage{
accuracy_tester_classification()
}
\value{
accuracy tester function for classification
}
\description{
\code{accuracy_tester_classification} returns a function that
compares the output of a neural network with the expected output
and returns true if they match and false if they don't
}
