% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_tester.R
\name{accuracy_tester_regression_abs}
\alias{accuracy_tester_regression_abs}
\title{Generate an absolute tester function for regression}
\usage{
accuracy_tester_regression_abs(maxAbsoluteDiviation)
}
\arguments{
\item{maxAbsoluteDiviation}{the maximal allowed absolute diviation from
the expected output}
}
\value{
absolute accuracy tester function for regression
}
\description{
\code{accuracy_tester_regression_abs} returns a function that
compares the output of a neural network with the expected output
and returns true if the absolute difference of the neural network
output to the expected output does not exceed \code{maxAbsoluteDiviation}.
Otherwise the returned function will return false
}
