% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_tester.R
\name{accuracy_tester_regression_rel}
\alias{accuracy_tester_regression_rel}
\title{Generate a relative tester function for regression}
\usage{
accuracy_tester_regression_rel(maxRelativeDiviation)
}
\arguments{
\item{maxRelativeDiviation}{the maximal allowed relative diviation from
the expected output}
}
\value{
relative accuracy tester function for regression
}
\description{
\code{accuracy_tester_regression_rel} returns a function that
compares the output of a neural network with the expected output
and returns true if the relative difference of the neural network
output to the expected output does not exceed \code{maxRelativeDiviation}.
Otherwise the returned function will return false
}
