% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{combineData}
\alias{combineData}
\title{Combine input and target lists}
\usage{
combineData(inputs, targets)
}
\arguments{
\item{inputs}{A list with the input data}

\item{targets}{A list with the target data}
}
\value{
The combined data list
}
\description{
Combines the given inputs and targets into one list that
can then be given to the trainer as a training / test data
list.
}
\examples{
inputs <-
  list(c(0.5, 0.5), c(0.1, 0.8))
targets <-
  list(2, 1)
combined <-
  combineData
    (inputs, targets)

}
