% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{NeuralNet}
\alias{NeuralNet}
\title{R6 Class Representing a Neural Network}
\description{
R6 Class Representing a Neural Network

R6 Class Representing a Neural Network
}
\examples{

## ------------------------------------------------
## Method `NeuralNet$plot`
## ------------------------------------------------

nn <- NeuralNet$new(c(2,4,4,2))

# will plot the Neural Network
nn$plot()

# will plot the Neural Network with normalized linewidth
nn$plot(standard.lwd = TRUE)

# will plot the Neural Network with only two colors,
# red for negative values, green for positive values
nn$plot(col.fct = function(x) {
    b <- (x < 0)
      v <- vector(length = length(b))
      v[b] <- "red"
      v[!b] <- "green"

      v
})
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NeuralNet$new()}}
\item \href{#method-calculate}{\code{NeuralNet$calculate()}}
\item \href{#method-plot}{\code{NeuralNet$plot()}}
\item \href{#method-clone}{\code{NeuralNet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuralNet$new(layers, activationfct = "ReLU", outputfct = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-calculate}{}}}
\subsection{Method \code{calculate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuralNet$calculate(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot an R6 NeuralNet using graphics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuralNet$plot(
  max.lwd = 5,
  standard.lwd = FALSE,
  col.fct = function(x) {     grDevices::hcl(x * 60 + 60) }
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max.lwd}}{numeric value representing the maximum linewidth, default is 5}

\item{\code{standard.lwd}}{boolean value wether or not all lines should be normalized}

\item{\code{col.fct}}{function taking a vecttor of values ranging from -1 to 1, returns a vector of colors to be used.
If \code{col.fct} is \code{NULL} all lines will be black}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{nn <- NeuralNet$new(c(2,4,4,2))

# will plot the Neural Network
nn$plot()

# will plot the Neural Network with normalized linewidth
nn$plot(standard.lwd = TRUE)

# will plot the Neural Network with only two colors,
# red for negative values, green for positive values
nn$plot(col.fct = function(x) {
    b <- (x < 0)
      v <- vector(length = length(b))
      v[b] <- "red"
      v[!b] <- "green"

      v
})
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuralNet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
