% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetwork-package.R
\docType{package}
\name{NeuralNetwork-package}
\alias{NeuralNetwork}
\alias{NeuralNetwork-package}
\title{NeuralNetwork: Creating, Training and Using Neural Networks}
\description{
Used to Create and Train Neural Networks
}
\author{
\strong{Maintainer}: Rico Sebastian \email{r.j.g.sebastian@gmail.com}

Authors:
\itemize{
  \item Dorothea Zeiler
  \item Karl Musulin
  \item Martin Koloseus
}

}
\keyword{internal}
