% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_momentum.R
\name{OptimizerMomentum}
\alias{OptimizerMomentum}
\title{Optimizer Stochastic Gradient Descent with Momentum}
\description{
This class is responsible for performing the
optimization method "Stochastic Gradient Descent"
with the momentum optimization
on a neural network for a given list of training data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerMomentum$new()}}
\item \href{#method-setLearningRate}{\code{OptimizerMomentum$setLearningRate()}}
\item \href{#method-setRegularizationRate}{\code{OptimizerMomentum$setRegularizationRate()}}
\item \href{#method-setMomentumTerm}{\code{OptimizerMomentum$setMomentumTerm()}}
\item \href{#method-getLearningRate}{\code{OptimizerMomentum$getLearningRate()}}
\item \href{#method-getRegularizationRate}{\code{OptimizerMomentum$getRegularizationRate()}}
\item \href{#method-getMomentumTerm}{\code{OptimizerMomentum$getMomentumTerm()}}
\item \href{#method-optim}{\code{OptimizerMomentum$optim()}}
\item \href{#method-reset}{\code{OptimizerMomentum$reset()}}
\item \href{#method-clone}{\code{OptimizerMomentum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new OptimizerMomentum object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$new(learning_rate, regularization_rate, momentum_term = 0.9)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{The learning rate used by the optimization method}

\item{\code{regularization_rate}}{The regularization rate used by the optimization method}

\item{\code{momentum_term}}{The momentum term used by the optimization method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLearningRate"></a>}}
\if{latex}{\out{\hypertarget{method-setLearningRate}{}}}
\subsection{Method \code{setLearningRate()}}{
Sets the learning rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$setLearningRate(learning_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{The new learning rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRegularizationRate"></a>}}
\if{latex}{\out{\hypertarget{method-setRegularizationRate}{}}}
\subsection{Method \code{setRegularizationRate()}}{
Sets the regularization rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$setRegularizationRate(regularization_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{regularization_rate}}{The new regularization rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMomentumTerm"></a>}}
\if{latex}{\out{\hypertarget{method-setMomentumTerm}{}}}
\subsection{Method \code{setMomentumTerm()}}{
Sets the momentum term
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$setMomentumTerm(momentum_term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{momentum_term}}{The new momentum term}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLearningRate"></a>}}
\if{latex}{\out{\hypertarget{method-getLearningRate}{}}}
\subsection{Method \code{getLearningRate()}}{
Returns the learning rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$getLearningRate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The learning rate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRegularizationRate"></a>}}
\if{latex}{\out{\hypertarget{method-getRegularizationRate}{}}}
\subsection{Method \code{getRegularizationRate()}}{
Returns the regularization rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$getRegularizationRate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The regularization rate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMomentumTerm"></a>}}
\if{latex}{\out{\hypertarget{method-getMomentumTerm}{}}}
\subsection{Method \code{getMomentumTerm()}}{
Returns the momentum term
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$getMomentumTerm()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The momentum term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optim"></a>}}
\if{latex}{\out{\hypertarget{method-optim}{}}}
\subsection{Method \code{optim()}}{
Performs the optimization algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$optim(neuralnet, training_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{neuralnet}}{The neural network to be trained}

\item{\code{training_data}}{The training data used for training the
network}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Deletes info from earlier optimization processes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
