% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_sgd.R
\name{OptimizerSGD}
\alias{OptimizerSGD}
\title{Optimizer Stochastic Gradient Descent}
\description{
This class is responsible for performing the
optimization algorithm on a neural network
for given training data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerSGD$new()}}
\item \href{#method-setLearningRate}{\code{OptimizerSGD$setLearningRate()}}
\item \href{#method-setLambda}{\code{OptimizerSGD$setLambda()}}
\item \href{#method-getLearningRate}{\code{OptimizerSGD$getLearningRate()}}
\item \href{#method-getLambda}{\code{OptimizerSGD$getLambda()}}
\item \href{#method-optim}{\code{OptimizerSGD$optim()}}
\item \href{#method-reset}{\code{OptimizerSGD$reset()}}
\item \href{#method-clone}{\code{OptimizerSGD$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$new(learning_rate, lambda)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLearningRate"></a>}}
\if{latex}{\out{\hypertarget{method-setLearningRate}{}}}
\subsection{Method \code{setLearningRate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$setLearningRate(learning_rate)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLambda"></a>}}
\if{latex}{\out{\hypertarget{method-setLambda}{}}}
\subsection{Method \code{setLambda()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$setLambda(lambda)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLearningRate"></a>}}
\if{latex}{\out{\hypertarget{method-getLearningRate}{}}}
\subsection{Method \code{getLearningRate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$getLearningRate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLambda"></a>}}
\if{latex}{\out{\hypertarget{method-getLambda}{}}}
\subsection{Method \code{getLambda()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$getLambda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optim"></a>}}
\if{latex}{\out{\hypertarget{method-optim}{}}}
\subsection{Method \code{optim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$optim(neuralnet, training_data, N = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
