% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpropagation.R
\name{sgdAlg}
\alias{sgdAlg}
\title{sgdAlg - Stochastic Gradient Descent}
\usage{
sgdAlg(neuralnet, training_data, learning_rate, lambda)
}
\arguments{
\item{neuralnet}{A R6 Neural Network that will be trained with the given data}

\item{training_data}{a data set used to train the Neural Network}

\item{learing_rate}{the learning rate to be used by the Algorithm}

\item{lamda}{a lambda to be used by the Algorithm}
}
\description{
sgdAlg implements an Algorithm for training a \code{?NeuralNet} Neural Network
using Stochastic Gradient Descent.
}
\seealso{
?NeuralNet
}
