% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{Trainer}
\alias{Trainer}
\title{Trainer class}
\description{
This class is responsible for managing the training
process of a given network for a given optimizer.
}
\seealso{
?NeuralNet
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Trainer$new()}}
\item \href{#method-setNeuralnet}{\code{Trainer$setNeuralnet()}}
\item \href{#method-setoptimizer}{\code{Trainer$setoptimizer()}}
\item \href{#method-setTrainingData}{\code{Trainer$setTrainingData()}}
\item \href{#method-setTestData}{\code{Trainer$setTestData()}}
\item \href{#method-getTrainingData}{\code{Trainer$getTrainingData()}}
\item \href{#method-getTestData}{\code{Trainer$getTestData()}}
\item \href{#method-getBestNeuralnet}{\code{Trainer$getBestNeuralnet()}}
\item \href{#method-swapWithBestNeuralnet}{\code{Trainer$swapWithBestNeuralnet()}}
\item \href{#method-seperate}{\code{Trainer$seperate()}}
\item \href{#method-test}{\code{Trainer$test()}}
\item \href{#method-train}{\code{Trainer$train()}}
\item \href{#method-reset}{\code{Trainer$reset()}}
\item \href{#method-clone}{\code{Trainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$new(
  neuralnet,
  optimizer,
  training_data = NULL,
  test_data = NULL,
  accuracy_measurement = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setNeuralnet"></a>}}
\if{latex}{\out{\hypertarget{method-setNeuralnet}{}}}
\subsection{Method \code{setNeuralnet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$setNeuralnet(neuralnet)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setoptimizer"></a>}}
\if{latex}{\out{\hypertarget{method-setoptimizer}{}}}
\subsection{Method \code{setoptimizer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$setoptimizer(optimizer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTrainingData"></a>}}
\if{latex}{\out{\hypertarget{method-setTrainingData}{}}}
\subsection{Method \code{setTrainingData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$setTrainingData(training_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTestData"></a>}}
\if{latex}{\out{\hypertarget{method-setTestData}{}}}
\subsection{Method \code{setTestData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$setTestData(test_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTrainingData"></a>}}
\if{latex}{\out{\hypertarget{method-getTrainingData}{}}}
\subsection{Method \code{getTrainingData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$getTrainingData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTestData"></a>}}
\if{latex}{\out{\hypertarget{method-getTestData}{}}}
\subsection{Method \code{getTestData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$getTestData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBestNeuralnet"></a>}}
\if{latex}{\out{\hypertarget{method-getBestNeuralnet}{}}}
\subsection{Method \code{getBestNeuralnet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$getBestNeuralnet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-swapWithBestNeuralnet"></a>}}
\if{latex}{\out{\hypertarget{method-swapWithBestNeuralnet}{}}}
\subsection{Method \code{swapWithBestNeuralnet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$swapWithBestNeuralnet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seperate"></a>}}
\if{latex}{\out{\hypertarget{method-seperate}{}}}
\subsection{Method \code{seperate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$seperate(data, test_percentage = 0.15)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test"></a>}}
\if{latex}{\out{\hypertarget{method-test}{}}}
\subsection{Method \code{test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$test(N)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
train trains the neural network of the trainer with the given optimizer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$train(
  epochs,
  training_per_epoch = Inf,
  use_early_stopping = F,
  es_test_frequency = 5000,
  es_test_size = 500,
  es_minimal_improvement = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epochs}}{integer; the number of epochs, the network will be trained}

\item{\code{training_per_epoch}}{integer; the amout of training data that will be used for an epoch}

\item{\code{use_early_stopping}}{logical; if true, early stopping is enabled}

\item{\code{es_test_frequency}}{integer; if early stopping is enabled,
test the performance of the network every \code{es_test_frequency} training sessions}

\item{\code{es_test_size}}{integer; if early stopping is enabled and the performance of
the network is to be evaluated, test the network with \code{es_test_size} test objects}

\item{\code{es_minimal_improvement}}{double; if early stopping is enabled, stop the training,
if the measured success ratio of the network didn't improve by at least
\code{es_minimal_improvement} in comparison to the best measured success ratio
for this network. Negative values are allowed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
