% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_sgd.R
\name{OptimizerSGD}
\alias{OptimizerSGD}
\title{Optimizer Stochastic Gradient Descent}
\description{
This class is responsible for performing the
optimization method "Stochastic Gradient Descent"
on a neural network for a given list of training data
}
\examples{
optimizer <- OptimizerSGD$new(0.0005, 0)
optimizer$setLearningRate(0.0001)
optimizer$setRegularizationRate(0.00001)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerSGD$new()}}
\item \href{#method-setLearningRate}{\code{OptimizerSGD$setLearningRate()}}
\item \href{#method-setRegularizationRate}{\code{OptimizerSGD$setRegularizationRate()}}
\item \href{#method-getLearningRate}{\code{OptimizerSGD$getLearningRate()}}
\item \href{#method-getRegularizationRate}{\code{OptimizerSGD$getRegularizationRate()}}
\item \href{#method-optim}{\code{OptimizerSGD$optim()}}
\item \href{#method-reset}{\code{OptimizerSGD$reset()}}
\item \href{#method-clone}{\code{OptimizerSGD$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new OptimiserSGD object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$new(learning_rate, regularization_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{The learning rate used by the optimization method}

\item{\code{regularization_rate}}{The regularization rate used by the optimization method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLearningRate"></a>}}
\if{latex}{\out{\hypertarget{method-setLearningRate}{}}}
\subsection{Method \code{setLearningRate()}}{
Sets the learning rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$setLearningRate(learning_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{The new learning rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRegularizationRate"></a>}}
\if{latex}{\out{\hypertarget{method-setRegularizationRate}{}}}
\subsection{Method \code{setRegularizationRate()}}{
Sets the regularization rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$setRegularizationRate(regularization_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{regularization_rate}}{The new regularization rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLearningRate"></a>}}
\if{latex}{\out{\hypertarget{method-getLearningRate}{}}}
\subsection{Method \code{getLearningRate()}}{
Returns the learning rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$getLearningRate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The learning rate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRegularizationRate"></a>}}
\if{latex}{\out{\hypertarget{method-getRegularizationRate}{}}}
\subsection{Method \code{getRegularizationRate()}}{
Returns the regularization rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$getRegularizationRate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The regularization rate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optim"></a>}}
\if{latex}{\out{\hypertarget{method-optim}{}}}
\subsection{Method \code{optim()}}{
Performs the optimization algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$optim(neuralnet, training_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{neuralnet}}{The neural network to be trained}

\item{\code{training_data}}{The training data used for training the
network}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Deletes info from earlier optimization processes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerSGD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
