% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{Trainer}
\alias{Trainer}
\title{train - Stochastic Gradient Descent}
\description{
train - Stochastic Gradient Descent

train - Stochastic Gradient Descent
}
\seealso{
?NeuralNet
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Trainer$new()}}
\item \href{#method-setTrainingData}{\code{Trainer$setTrainingData()}}
\item \href{#method-setTestData}{\code{Trainer$setTestData()}}
\item \href{#method-getTrainingData}{\code{Trainer$getTrainingData()}}
\item \href{#method-getTestData}{\code{Trainer$getTestData()}}
\item \href{#method-test}{\code{Trainer$test()}}
\item \href{#method-train}{\code{Trainer$train()}}
\item \href{#method-clone}{\code{Trainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
train implements an Algorithm for training a \code{?NeuralNet} Neural Network
using Stochastic Gradient Descent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$new(optimiser, training_data, test_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTrainingData"></a>}}
\if{latex}{\out{\hypertarget{method-setTrainingData}{}}}
\subsection{Method \code{setTrainingData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$setTrainingData(training_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTestData"></a>}}
\if{latex}{\out{\hypertarget{method-setTestData}{}}}
\subsection{Method \code{setTestData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$setTestData(test_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTrainingData"></a>}}
\if{latex}{\out{\hypertarget{method-getTrainingData}{}}}
\subsection{Method \code{getTrainingData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$getTrainingData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTestData"></a>}}
\if{latex}{\out{\hypertarget{method-getTestData}{}}}
\subsection{Method \code{getTestData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$getTestData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test"></a>}}
\if{latex}{\out{\hypertarget{method-test}{}}}
\subsection{Method \code{test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$test(neuralnet, N)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$train(neuralnet, learning_rate, lambda, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{neuralnet}}{A R6 Neural Network that will be trained with the given data}

\item{\code{training_data}}{a data set used to train the Neural Network}

\item{\code{learing_rate}}{the learning rate to be used by the Algorithm}

\item{\code{lamda}}{a lambda to be used by the Algorithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
